+++ score/src/main/java/org/shareio/backend/core/model/ReviewSnapshot.java
@@ -4,7 +4,7 @@ import org.shareio.backend.core.model.vo.ReviewId;
 
 import java.time.LocalDateTime;
 
-public record ReviewSnapshot(ReviewId reviewId, Float value, LocalDateTime date) {
+public record ReviewSnapshot(ReviewId reviewId, Double value, LocalDateTime date) {
     public ReviewSnapshot(Review review) {
         this(review.getReviewId(),  review.getValue(), review.getDate());
     }
